Add files via upload
in this file, is some of the handwritten code, that i wrote from memory using the previous knowledge of deep learning models. Albeit, it's not fully complete, however, it does show my understanding and grasp of the coding concepts in order to create/manipulate data to make it useful for coding.