Add files via upload
" In IS-160, the first chapters we learned about deep model layers and how to manually calculate using forward propagation. which issued to find the values of the hidden layers and further used to find the values of the output layer. by learning how to do the manual calculation is key in understanding how the code is able to find the values in such a short compute. "