Add files via upload
"in this file, it contains the lab for the stable diffusion work, or aka prompt engineering. where we ran the stable diffusion code, which allows us to use the program in order to practice prompt engineering. In this case, it helps understand the value of the prompt and how much affect it can have on the results, since the more or less detail your prompt has the more detailed the image generated will be."