Add files via upload
In this lab, we started to mess with reinforcement learning, where we must focus on an agent, states, environment, rewards. where the agent must interact with the environment in order to achieve its reward. depending on how well the agent does determines the reward. and each state is a probability in which the agents can result in.